fix(preview): stabilize huge workspace metrics, rendering, and sticky header

+++ Tests/DevProjex.Tests.Integration/EffectiveIgnoreOptionDirectMatchCountMatrixIntegrationTests.cs
@@ -258,7 +258,7 @@ public sealed class EffectiveIgnoreOptionDirectMatchCountMatrixIntegrationTests
 			.Root;
 	}
 
-	private static int CountAsciiTreeLines(string rootPath, TreeNodeDescriptor root)
+	private static long CountAsciiTreeLines(string rootPath, TreeNodeDescriptor root)
 	{
 		var treeText = new TreeExportService().BuildFullTree(rootPath, root, TreeTextFormat.Ascii);
 		return ExportOutputMetricsCalculator.FromText(treeText).Lines;
